requirements.txt
pandas
numpy
scipy
scikit-learn
statsmodels
pyyaml
matplotlib
seaborn
